requirements.txt
Flask==1.1.2
Pillow==7.2.0
tensorflow==2.5.0
numpy==1.19.2
